chore(tests): rename `test_pep695_new_syntax.py` to `pep695_new_syntax.py`
