+++ tests/unit/container/pep695_new_syntax.py
